refactor: convert config file to ts
